Change the add-on to use modules where possible
Instead of injecting several scripts, use JavaScript modules and
imports/exports to make it cleaner.

There are a few locations where we still require several scripts:

- The content script cannot use modules as far as I know (although I
  might be wrong). Dynamic imports are reportedly working on Firefox and
  could be used in the future.
- As a consequence, the functions in transformations.js cannot be
  exported as part of a module, and they are globally injected for the
  option UI as well.
- The highlight.js library can be used as a module generally, but I
  don't think this can be done for the content script. I also couldn't
  find a way to do it for the option UI from the version we embed. This
  might change in the future if I find a way to do it.
