+++ options/ui.html
@@ -82,11 +82,7 @@
         Click to reset all options to their default values.
       </p>
     <script src="/hljs/highlight.min.js"></script>
-    <script src="defaults.js"></script>
-    <script src="fill-css-list.js"></script>
     <script src="/scripts/transformations.js"></script>
-    <script src="preview.js"></script>
-    <script src="save-restore.js"></script>
-    <script src="listeners.js"></script>
+    <script type="module" src="listeners.js"></script>
   </body>
 </html>
